Limit the number of iterations in Macros.remove
There is a bug in rpm 4.16 which makes %echo expansion fail which made
Macros.remove() loop infinitely. This also applies to other parametrized
builtin macros. Even though we do not make use of such macros, such
behaviour is undesirable due to the general target use. Limit the number of
iterations to workaround this problem.

Signed-off-by: František Nečas <fnecas@redhat.com>