Handle sources referenced from shell expansions in tag values
Sources referenced from shell expansions in tag values, for example
"Summary: %(cat %{SOURCE1})", are required even for non-build parse
and can't be faked with empty files, because empty tag values are not
allowed in a spec file.

If requested, fake such sources with a dummy value and mark
the corresponding `Specfile` instance as tainted after doing so.

Also rename the `ignore_missing_includes` option to a more general
`force_parse`.

Signed-off-by: Nikola Forró <nforro@redhat.com>