Make context managers better
This commit introduces a new ContextManager decorator that builds on top
of contextlib.contextmanager and effectively makes context managers
a singleton. Any nested context of the same kind will be the same
as the outermost context of that particular kind.
This makes it possible, for example, to set properties that internally
use context managers inside an existing context, or in general mix
and nest context managers almost freely.
There are still instances of conflicting context managers, for example
macro_definitions() and sections(), that, when combined, will overwrite
each other's underlying data. All such instances will be documented.

Signed-off-by: Nikola Forró <nforro@redhat.com>