Fix parsing of tags
When parsing tags, if there is a conditional macro expansion on a line,
the line is split into prefix, content and suffix. If content is a tag,
a `Tag` object is created and prefix and suffix are stored within it.
But in case the content is not a tag, prefix and suffix are lost.
So when reconstructing section data later, the reconstructed line is not
the same and that can cause all kinds of issues.

Fix that.

Signed-off-by: Nikola Forró <nforro@redhat.com>