changelog: add guess_packager() function
Add a `specfile.changelog.guess_packager()` with similar behavior to
rpmdev-packager. This makes it so specfile doesn't need to shell out to
external binaries that may or may not be installed.

There's a docstring and unit tests that cover most of the potential
cases. I might've gone a bit overboard :).

- add test_guess_packager_empty

- use Macros.expand() helper