Make modifiable entity groups greedy
This effectively makes the first group most greedy, rather than the
last. So, with the following in a spec file:

```
%global apiver 2.15
%global patchver 3

Version: %{apiver}.%{patchver}
```

calling `spec.update_tag("Version", "2.16.0")`, the `apiver` macro
gets assigned the bigger portion of the version string, resulting in:

```
%global apiver 2.16
%global patchver 0
```

rather than:

```
%global apiver 2
%global patchver 16.0
```

There are cases (in Fedora spec files) where the previous logic
makes more sense, but only a handful of them, and none of them
using packit.
With the current implementation, it's one or the other, and I believe
this makes more sense.

Signed-off-by: Nikola Forró <nforro@redhat.com>