Recognize all macro definitions in hash-style comments
Macro definitions can be commented out by replacing '%' with '#',
for example `%global debug 1` with `#global debug 1`, to avoid
`Macro expanded in comment` warning. However, the same can be achieved
also by escaping the '%' in the comment, i.e. `#%%global debug 1`,
and even if the '%' is not escaped, it is still parseable, so we should
be able to parse it as well.

Signed-off-by: Nikola Forró <nforro@redhat.com>