Explicitly invalidate the global parse hash in `SpecParser` constructor
Python can reuse id of objects after they are garbage-collected, however
the global parse hash contains an id of `SpecParser` instance in order
to force parsing on context switches, and when a different instance has
the same id, it has no way to detect that.

Explicitly invalidate the global parse hash when a `SpecParser` instance
is created to prevent this issue.

Signed-off-by: Nikola Forró <nforro@redhat.com>