Add support for (open)SUSE-style changelog entries
These changelog entries are created using the vc utility part of of
obs-build (https://github.com/openSUSE/obs-build/blob/master/vc) and are saved
in a separate .changes file.

This commit creates an abstract base class ChangelogEntryBase with all
properties & methods of the ChangelogEntry class and adds the child classes
ChangelogEntry and SUSEChangelogEntry which implement it.